features/add middleware validate_body
