+++ schemas/contacts.js
@@ -1,0 +1,20 @@
+const Joi = require("joi");
+
+const addSchema = Joi.object({
+  id: Joi.string().required(),
+  name: Joi.string().required(),
+  phone: Joi.string().required(),
+  email: Joi.string().required(),
+});
+
+const updateSchema = Joi.object({
+  id: Joi.string(),
+  name: Joi.string(),
+  phone: Joi.string(),
+  email: Joi.string(),
+});
+
+module.exports = {
+  addSchema,
+  updateSchema,
+};
